automatic synch
new tasks that automatically synchronize esp and ds3231 times via callbacks and seperate tasks. Some shinanigens to create FreeRtos Tasks in our classes and use the instance of the class itself --> using a static "This" version to solve these problems, because FreeRtos Task must be static functions, and static functions do not have an instance bound to them!